quickpick - move Themable to platform
