+++ src/vs/platform/theme/common/themeService.ts
@@ -5,7 +5,7 @@
 
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { Color } from 'vs/base/common/color';
-import { IDisposable, toDisposable } from 'vs/base/common/lifecycle';
+import { IDisposable, toDisposable, Disposable } from 'vs/base/common/lifecycle';
 import * as platform from 'vs/platform/registry/common/platform';
 import { ColorIdentifier } from 'vs/platform/theme/common/colorRegistry';
 import { Event, Emitter } from 'vs/base/common/event';
@@ -190,4 +190,42 @@ platform.Registry.add(Extensions.ThemingContribution, themingRegistry);
 
 export function registerThemingParticipant(participant: IThemingParticipant): IDisposable {
 	return themingRegistry.onColorThemeChange(participant);
+}
+
+/**
+ * Utility base class for all themable components.
+ */
+export class Themable extends Disposable {
+	protected theme: IColorTheme;
+
+	constructor(
+		protected themeService: IThemeService
+	) {
+		super();
+
+		this.theme = themeService.getColorTheme();
+
+		// Hook up to theme changes
+		this._register(this.themeService.onDidColorThemeChange(theme => this.onThemeChange(theme)));
+	}
+
+	protected onThemeChange(theme: IColorTheme): void {
+		this.theme = theme;
+
+		this.updateStyles();
+	}
+
+	protected updateStyles(): void {
+		// Subclasses to override
+	}
+
+	protected getColor(id: string, modify?: (color: Color, theme: IColorTheme) => Color): string | null {
+		let color = this.theme.getColor(id);
+
+		if (color && modify) {
+			color = modify(color, this.theme);
+		}
+
+		return color ? color.toString() : null;
+	}
 }
